improve Bezier
